fix(api): radiology report create compiles again, and clippy is clean
16bddb2 (demo dataset seeding) read `report.report_id` in the duplicate
branch after `report` had been moved into the repository call, so
medichain-api did not compile (E0382) and `cargo test` could not run.
Keep the id before the move.

Also drop an unnecessary `u64 as u64` cast in the fall-risk repository,
the one clippy warning on the current toolchain.

+++ api/src/repositories/postgres/fall_risk_assessment.rs
@@ -124,7 +124,7 @@ impl FallRiskAssessmentRepository for PgFallRiskAssessmentRepository {
             .fetch_all(&self.pool)
             .await?;
 
-        Ok(PaginatedResult::new(assessments, total as u64, &pagination))
+        Ok(PaginatedResult::new(assessments, total, &pagination))
     }
 
     async fn get_latest_by_patient(
